Chapter05/requirements.txt
aiodns==1.1.1
aiohttp==2.0.7
amqp==2.1.4
async-timeout==1.2.0
billiard==3.5.0.2
cchardet==2.0.0
celery==4.0.2
chardet==3.0.2
eyeD3==0.7.11
id3reader==1.53.20070415
joblib==0.11
kombu==4.0.2
multidict==2.1.4
numpy==1.12.1
olefile==0.44
Pillow==4.1.0
pkg-resources==0.0.0
pycares==2.1.1
pymp-pypi==0.4.0
pytz==2017.2
redis==2.10.5
vine==1.1.3
yarl==0.10.0
